+++ src/SlimFaas/Database/IDatabaseService.cs
@@ -7,6 +7,6 @@ public interface IDatabaseService
     Task HashSetAsync(string key, IDictionary<string, string> values);
     Task<IDictionary<string, string>> HashGetAllAsync(string key);
     Task ListLeftPushAsync(string key, string field);
-    Task<IList<string>> ListRightPopAsync(string key, long count = 1);
+    Task<IList<string>> ListRightPopAsync(string key, int count = 1);
     Task<long> ListLengthAsync(string key);
 }
